Move onLengthChange function
