+++ packages/gafl-webapp-service/src/pages/licence-details/licence-to-start/update-transaction.js
@@ -1,10 +1,8 @@
 import moment from 'moment'
-import { LICENCE_TO_START, LICENCE_START_TIME } from '../../../uri.js'
+import { LICENCE_START_TIME, LICENCE_TO_START } from '../../../uri.js'
 import { ageConcessionHelper } from '../../../processors/concession-helper.js'
-import { SERVICE_LOCAL_TIME } from '@defra-fish/business-rules-lib'
 import { cacheDateFormat } from '../../../processors/date-and-time-display.js'
-import { isPhysical } from '../../../processors/licence-type-display.js'
-import { HOW_CONTACTED } from '../../../processors/mapping-constants.js'
+import { onLengthChange } from '../licence-length/update-transaction'
 
 /**
  * Transfer the validate page object
@@ -14,31 +12,6 @@ import { HOW_CONTACTED } from '../../../processors/mapping-constants.js'
 export const licenceToStart = {
   AFTER_PAYMENT: 'after-payment',
   ANOTHER_DATE: 'another-date'
-}
-
-export const onLengthChange = permission => {
-  // If the licence start date has be chosen as today, and the licence is changed to a 12 month
-  // then set the start after payment flag
-  if (
-    permission.licenceLength === '12M' &&
-    moment(permission.licenceStartDate, cacheDateFormat)
-      .tz(SERVICE_LOCAL_TIME)
-      .isSame(moment(), 'day')
-  ) {
-    permission.licenceToStart = licenceToStart.AFTER_PAYMENT
-    permission.licenceStartTime = null
-  }
-
-  // If we have set the licence type to physical change the method of contact from 'none' to 'letter' and vice-versa
-  if (isPhysical(permission) && permission?.licensee?.preferredMethodOfConfirmation === HOW_CONTACTED.none) {
-    permission.licensee.preferredMethodOfConfirmation = HOW_CONTACTED.letter
-    permission.licensee.preferredMethodOfReminder = HOW_CONTACTED.letter
-  }
-
-  if (!isPhysical(permission) && permission?.licensee?.preferredMethodOfConfirmation === HOW_CONTACTED.letter) {
-    permission.licensee.preferredMethodOfConfirmation = HOW_CONTACTED.none
-    permission.licensee.preferredMethodOfReminder = HOW_CONTACTED.none
-  }
 }
 
 export default async request => {
