+++ packages/gafl-webapp-service/src/pages/licence-details/licence-type/update-transaction.js
@@ -1,7 +1,7 @@
 import { LICENCE_TYPE } from '../../../uri.js'
 import * as constants from '../../../processors/mapping-constants.js'
 import { licenseTypes } from './route.js'
-import { onLengthChange } from '../licence-to-start/update-transaction.js'
+import { onLengthChange } from '../licence-length/update-transaction'
 
 /**
  * Transfer the validate page object
